sql: fix 2-param version of setval to not set isCalled
Release note (bug fix, backward-incompatible change): The 2-parameter
setval builtin previously caused the sequence to increment incorrectly
one extra time. This is fixed now. If you do need the sequence to
increment, then use `setval(seq, val, true)`.

+++ pkg/ccl/importccl/import_table_creation.go
@@ -395,7 +395,7 @@ func (so *importSequenceOperators) GetLatestValueInSessionForSequenceByID(
 
 // SetSequenceValueByID implements the tree.SequenceOperators interface.
 func (so *importSequenceOperators) SetSequenceValueByID(
-	ctx context.Context, seqID int64, newVal int64, isCalled bool,
+	ctx context.Context, seqID uint32, newVal int64, isCalled bool,
 ) error {
 	return errSequenceOperators
 }
